refactored tests, warning disappeared
